added module high five and set phase passes rspec
