+++ faculty.rb
@@ -6,7 +6,11 @@ class Faculty
     @phase = 3
     @age = options.fetch(:age, 0)
     @name = options.fetch(:name, "")
-    # @target_raise =
+    # @target_raise = self::TARGET_RAISE
   end
 
+  def set_phase(num)
+    @phase = num
+    "Cool, I've always wanted to teach phase #{num}!"
+  end
 end
