+++ high_five.rb
@@ -1,0 +1,5 @@
+module High_Five
+  def offer_high_five
+    "High five!"
+  end
+end
