+++ student.rb
@@ -1,3 +1,5 @@
+require_relative 'high_five'
+
 class Student
   attr_reader :age, :phase
   attr_accessor :name
@@ -8,9 +10,8 @@ class Student
     @name = options.fetch(:name, "")
   end
 
-  def offer_high_five
-    "High five!"
-  end
+  include High_Five
+
 
   def set_phase(num)
     response = ""
